python/requirements.txt
Flask==3.0.0
globalpayments.api==2.0.4
gunicorn==21.2.0
python-dotenv==1.0.0
